Separate meeting views from controller logic
