+++ src/services/meetings/meeting.routes.ts
@@ -1,59 +1,11 @@
 import express from "express";
 import { Meeting } from "./meeting.model.js";
 import { AuthenticatedRequest } from "../../middlewares/auth.js";
+import { getAllMeetings, getMeetingStats } from "./meeting.controller.js";
 
 export const router = express.Router();
 
-// GET all meetings for user
-router.get("/", async (req: AuthenticatedRequest, res) => {
-  try {
-    const meetings = await Meeting.find();
-    res.json({
-      total: meetings.length,
-      limit: req.query.limit,
-      page: req.query.page,
-      data: meetings,
-    });
-  } catch (err) {
-    res.status(500).json({ message: (err as Error).message });
-  }
-});
-
-// TODO: implement other endpoints
-
-router.get("/stats", async (req: AuthenticatedRequest, res) => {
-  try {
-    // TODO: get statistics from the database
-    const stats = {
-      generalStats: {
-        totalMeetings: 100,
-        averageParticipants: 4.75,
-        totalParticipants: 475,
-        shortestMeeting: 15,
-        longestMeeting: 120,
-        averageDuration: 45.3,
-      },
-      topParticipants: [
-        { participant: "John Doe", meetingCount: 20 },
-        { participant: "Jane Smith", meetingCount: 18 },
-        { participant: "Bob Johnson", meetingCount: 15 },
-        { participant: "Alice Brown", meetingCount: 12 },
-        { participant: "Charlie Davis", meetingCount: 10 },
-      ],
-      meetingsByDayOfWeek: [
-        { dayOfWeek: 1, count: 10 },
-        { dayOfWeek: 2, count: 22 },
-        { dayOfWeek: 3, count: 25 },
-        { dayOfWeek: 4, count: 20 },
-        { dayOfWeek: 5, count: 18 },
-        { dayOfWeek: 6, count: 5 },
-        { dayOfWeek: 7, count: 0 },
-      ],
-    };
-    res.json(stats);
-  } catch (err) {
-    res.status(500).json({ message: (err as Error).message });
-  }
-});
+router.get("/", getAllMeetings);
+router.get("/stats", getMeetingStats);
 
 export { router as meetingRoutes };
